+++ packages/antd-ext/components/ButtonExt/Mix.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import ButtonExt from './ButtonExt';
 import type { ButtonExtProps } from './ButtonExt';
-import { Theme as ThemeHermes, customStyle as customStyleHermes } from './ThemeHermes';
+import { customStyle as customStyleHermes } from './ThemeHermes';
 import type { Theme } from '../ConfigProviderExt/context';
 import useMapTheme from '../utils/useMapTheme';
+import ThemeWrapper from '../theme/ThemeWrapper';
 
 export interface ButtonExtMixinProps extends ButtonExtProps {
   theme?: Theme;
@@ -11,19 +12,16 @@ export interface ButtonExtMixinProps extends ButtonExtProps {
 
 export default function Mixin(props: ButtonExtMixinProps) {
   const { className, theme, ...restProps } = props;
-  const { classes, ThemeWrapper, tokenExt } = useMapTheme({
+  const { classes, themeConfig } = useMapTheme({
     className,
     theme,
-    themeWrap: {
-      hermes: ThemeHermes,
-    },
     emotioncss: {
       hermes: customStyleHermes,
     },
   });
 
   return (
-    <ThemeWrapper tokenExt={tokenExt}>
+    <ThemeWrapper autoInsertSpaceInButton={theme !== 'hermes'} theme={themeConfig} type="Button">
       <ButtonExt className={classes} {...restProps} />
     </ThemeWrapper>
   );