+++ packages/antd-ext/examples/TableExt/base/TableExtBackEnd.tsx
@@ -160,6 +160,19 @@ export default function TableExtBackEnd(props: {
               }
             : undefined
         }
+        summary={(currentData, fetching) => {
+          return (
+            <TableExt.Summary.Row>
+              <TableExt.Summary.Cell index={0}>
+                {fetching ? <TableExt.SkeletonItem /> : 'Summary'}
+              </TableExt.Summary.Cell>
+              <TableExt.Summary.Cell colSpan={3} index={1}>
+                {fetching ? <TableExt.SkeletonItem /> : 'This is a summary content'}
+              </TableExt.Summary.Cell>
+            </TableExt.Summary.Row>
+          );
+        }}
+        footer={() => 'Footer'}
       />
     </Card>
   );